style(rust): apply clippy autofixes
Mechanical cleanups across the agent harness and lib: sort_by -> sort_by_key,
redundant closures removed, format!("literal") -> to_string(), manual prefix
stripping -> strip_prefix. No behavioral change; all 71 tests pass.

+++ src-tauri/src/memory.rs
@@ -343,8 +343,8 @@ fn parse_iso_ms(s: &str) -> Option<i64> {
     let era = if y >= 0 { y } else { y - 399 } / 400;
     let yoe = (y - era * 400) as u64;
     let doy =
-        ((153 * (if month > 2 { month - 3 } else { month + 9 }) + 2) / 5 + day as i64 - 1) as u64;
-    let doe = yoe as u64 * 365 + yoe / 4 - yoe / 100 + doy;
+        ((153 * (if month > 2 { month - 3 } else { month + 9 }) + 2) / 5 + day - 1) as u64;
+    let doe = yoe * 365 + yoe / 4 - yoe / 100 + doy;
     let days = era * 146_097 + doe as i64 - 719_468;
     Some((days * 86_400 + hour * 3600 + min * 60 + sec) * 1000)
 }
